+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Baywatch default parameters added: `CB_BAYWATCH_ENABLED` and`CB_BAYWATCH_EXTERN_LOCATION`
 - Logs are saved via lospout
+- TLS client certificate needed by Cloudbreak is generated with `cbd generate`
 
 ### Removed